manatal_service/requirements.txt
faiss_cpu
fastapi
flashtext
numpy
pydantic
pydantic[email]
python-dotenv
requests
sentence_transformers
spacy
streamlit
tqdm
uvicorn

